Switch to Obfuscate callback for $DatabasePassword/$LDAPPassword configs
Previously we checked if the config name looks like password and
explicitly excluded MinimumPasswordLength, which didn't scale well: we
would have to exclude the new added $DisablePasswordForAuthToken in
various places to not obfuscate it.

This commit simplifies the logic: all configs that require obfuscation
need to set up Obfuscate callback accordingly.